+++ src/utils.ts
@@ -1,5 +1,6 @@
 import * as vscode from "vscode";
 
+//Method To show Progeress Notification in VSCODE
 function showProgress(title: string, fun: Function): Thenable<any> {
   return vscode.window.withProgress(
     {
@@ -11,10 +12,11 @@ function showProgress(title: string, fun: Function): Thenable<any> {
     }
   );
 }
+//Method To show Notification in VSCODE
 function showNotification(message: string) {
   var noti = vscode.window.showInformationMessage(message);
 }
-
+//Method To show error in VSCODE
 function showError(message: string) {
   vscode.window.showErrorMessage(message);
 }