+++ CHANGELOG.md
@@ -9,3 +9,15 @@ All notable changes to the "adb-qr" extension will be documented in this file.
 ## [0.0.2]
 
 - Bug Fixes
+
+## [0.0.3]
+
+- Minor Changes
+
+## [0.1.0]
+
+- Added Progress Indicator in place of static notifications
+
+## [0.1.1]
+
+- Fixed connect with paired device running in background